+++ src/datadoc/backend/src/model_validation.py
@@ -15,13 +15,13 @@ from datadoc.backend.src.utility.constants import DATE_VALIDATION_MESSAGE
 from datadoc.backend.src.utility.constants import NUM_OBLIGATORY_DATASET_FIELDS
 from datadoc.backend.src.utility.constants import NUM_OBLIGATORY_VARIABLES_FIELDS
 from datadoc.backend.src.utility.constants import OBLIGATORY_METADATA_WARNING
-from datadoc.backend.src.utils import get_missing_obligatory_dataset_fields
-from datadoc.backend.src.utils import get_missing_obligatory_variables_fields
-from datadoc.backend.src.utils import get_timestamp_now
-from datadoc.backend.src.utils import incorrect_date_order
-from datadoc.backend.src.utils import num_obligatory_dataset_fields_completed
-from datadoc.backend.src.utils import num_obligatory_variables_fields_completed
-from datadoc.backend.src.utils import set_variables_inherit_from_dataset
+from datadoc.backend.src.utility.utils import get_missing_obligatory_dataset_fields
+from datadoc.backend.src.utility.utils import get_missing_obligatory_variables_fields
+from datadoc.backend.src.utility.utils import get_timestamp_now
+from datadoc.backend.src.utility.utils import incorrect_date_order
+from datadoc.backend.src.utility.utils import num_obligatory_dataset_fields_completed
+from datadoc.backend.src.utility.utils import num_obligatory_variables_fields_completed
+from datadoc.backend.src.utility.utils import set_variables_inherit_from_dataset
 
 if TYPE_CHECKING:
     from datetime import datetime
