+++ src/datadoc/backend/src/utility/utils.py
