+++ src/datadoc/backend/tests/test_utils.py
@@ -6,9 +6,9 @@ import pytest
 from cloudpathlib.local import LocalGSClient
 from cloudpathlib.local import LocalGSPath
 
-from datadoc.backend.src.utils import calculate_percentage
-from datadoc.backend.src.utils import incorrect_date_order
-from datadoc.backend.src.utils import normalize_path
+from datadoc.backend.src.utility.utils import calculate_percentage
+from datadoc.backend.src.utility.utils import incorrect_date_order
+from datadoc.backend.src.utility.utils import normalize_path
 from datadoc.backend.tests.constants import DATADOC_METADATA_MODULE_UTILS
 from datadoc.backend.tests.constants import TEST_BUCKET_PARQUET_FILEPATH
 from datadoc.backend.tests.constants import TEST_PARQUET_FILEPATH
